+++ components/date-picker/DatePicker.js
@@ -1,12 +1,82 @@
 import React from 'react'
+import PropTypes from 'prop-types'
 
 import BasePicker from './BasePicker'
 import DatePanel from './DatePanel'
 import DateRangePanel from './DateRangePanel'
 import WeekRangePanel from './WeekRangePanel'
 import Provider from '../context'
-
+import { getPRCDate, deconstructDate } from './util'
 class DatePicker extends BasePicker {
+  static propTypes = {
+    altCalendar: PropTypes.array,
+    dateMarkRender: PropTypes.func,
+    altCalendarPreset: PropTypes.oneOf(['zh-CN', 'id-ID']),
+    dateMarkPreset: PropTypes.oneOf(['zh-CN'])
+  }
+  constructor (props) {
+    super(props)
+    this.altCalendarPresetData = {}
+    this.dateMarkPresetData = {}
+    this._getLunarPresetData()
+    this._getMarkPresetData()
+  }
+  // 获取预置数据
+  _getLunarPresetData () {
+    const {altCalendarPreset, altCalendar} = this.props
+    const allPRCDate = {}
+    if (['zh-CN', 'id-ID'].includes(altCalendarPreset)) {
+      const _urlKey = altCalendarPreset === 'zh-CN' ? 'PRCLunar' : 'IndiaHoliday'
+      getPRCDate(_urlKey).then(res => {
+        Object.keys(res.data).forEach(key => {
+          let oneYear = {}
+          res.data[key][_urlKey].forEach(item => {
+            Object.assign(oneYear, {
+              [item.date.replace(/-/g, '/')]: {
+                ...item,
+                highlight: altCalendarPreset === 'id-ID'
+              }
+            })
+          })
+          Object.assign(allPRCDate, oneYear)
+        })
+        this.altCalendarPresetData = altCalendar ? this._altCalendarData(allPRCDate) : allPRCDate
+      })
+    } else {
+      this.altCalendarPresetData = altCalendar ? this._altCalendarData(allPRCDate) : {}
+    }
+  }
+  // 获取预置数据
+  _getMarkPresetData () {
+    const {altCalendarPreset, dateMarkPreset} = this.props
+    if (altCalendarPreset && altCalendarPreset !== 'zh-CN') {
+      return
+    }
+    dateMarkPreset === 'zh-CN' && getPRCDate('PRCHoliday').then(res => {
+      const allPRCDate = {}
+      Object.keys(res.data).forEach(key => {
+        Object.keys(res.data[key].PRCHoliday).forEach(elkey => {
+          allPRCDate[elkey.replace(/-/g, '/')] = res.data[key].PRCHoliday[elkey] === '1'
+            ? <span className='hi-datepicker__text——holiday hi-datepicker__text——holiday--rest'>休</span>
+            : <span className='hi-datepicker__text——holiday hi-datepicker__text——holiday--work'>班</span>
+        })
+      })
+      this.dateMarkPresetData = allPRCDate
+    })
+  }
+  // 合并用户自定义的日期信息作为presetData
+  _altCalendarData = (allPRCDate) => {
+    const allData = {}
+    this.props.altCalendar.length > 0 && this.props.altCalendar.forEach(item => {
+      const dateInfo = deconstructDate(item.date)
+      if (!Number.isNaN(dateInfo.year)) {
+        Object.assign(allData, {
+          [dateInfo.year + '/' + dateInfo.month + '/' + dateInfo.day]: item
+        })
+      }
+    })
+    return Object.assign(allPRCDate, allData)
+  }
   initPanel (state, props) {
     let component = null
     let d = state.date
@@ -18,6 +88,8 @@ class DatePicker extends BasePicker {
         component = (
           <DatePanel
             {...props}
+            altCalendarPresetData={this.altCalendarPresetData}
+            dateMarkPresetData={this.dateMarkPresetData}
             date={state.date}
             format={this.state.format}
             onPick={this.onPick.bind(this)}
@@ -29,6 +101,8 @@ class DatePicker extends BasePicker {
         component = (
           <DatePanel
             {...props}
+            altCalendarPresetData={this.altCalendarPresetData}
+            dateMarkPresetData={this.dateMarkPresetData}
             format={this.state.format}
             date={state.date}
             onPick={this.onPick.bind(this)}
@@ -40,6 +114,8 @@ class DatePicker extends BasePicker {
         component = (
           <DateRangePanel
             {...props}
+            altCalendarPresetData={this.altCalendarPresetData}
+            dateMarkPresetData={this.dateMarkPresetData}
             format={this.state.format}
             date={d}
             onPick={this.onPick.bind(this)}